Move fixtures to a catalog
