+++ tests/e2e/entity/delete.test.ts
@@ -2,7 +2,7 @@ import { afterAll, beforeAll, describe, expect, test, vi } from 'vitest';
 
 import { NotFoundError } from '@lib/utils';
 
-import { MockEntityManager, TEST_TABLE_NAME, TestTableManager } from '../../fixtures';
+import { MockEntityManager, TEST_TABLE_NAME, TestTableManager } from '../../fixtures/TestTable';
 import { mockEntityFactory } from '../mockEntityFactory';
 
 describe('EntityManager.delete', () => {
