+++ tests/unit/decorators/helpers/prefixSuffix.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, test, vi } from 'vitest';
 import { prefix, suffix } from '@lib/decorators/helpers/prefixSuffix';
 import Dynamode from '@lib/dynamode/index';
 
-import { MockEntity, mockInstance } from '../../../fixtures';
+import { MockEntity, mockInstance } from '../../../fixtures/TestTable';
 
 describe('Decorators', () => {
   describe('prefix', async () => {
