+++ tests/unit/scan/index.test.ts
@@ -7,7 +7,7 @@ import Scan from '@lib/scan';
 import { Metadata } from '@lib/table/types';
 import { BASE_OPERATOR } from '@lib/utils';
 
-import { MockEntity, MockEntityManager, mockInstance, TEST_TABLE_NAME } from '../../fixtures';
+import { MockEntity, MockEntityManager, mockInstance, TEST_TABLE_NAME } from '../../fixtures/TestTable';
 
 vi.mock('@lib/utils/ExpressionBuilder', () => {
   const ExpressionBuilder = vi.fn(() => ({
